Game speed

+++ project/Game/Score.cpp
@@ -19,8 +19,8 @@ Score::Score()
     scorePos = new SDL_Rect;
     scorePos->x = 10;
     scorePos->y = 4;
-    scorePos->w = 147*0.2;
-    scorePos->h = 222*0.2;
+    scorePos->w = static_cast<int>(147*0.2);
+    scorePos->h = static_cast<int>(222*0.2);
     word[0].SetText("SCORE");
     word[0].SetPosition(40,6);
     word[0].ReduceSize(0.8);
